mock hooks inline
